move news lookupSchool hook to main hooks
